feat: add auth module into user module, modify auth files
