+++ .env.example
@@ -6,3 +6,5 @@ DATABASE_SCHEMA=momentrip
 DATABASE_HOST=localhost
 DATABASE_PORT=3306
 
+KAKAO_KEY=9c3b763915836894e84e16dd28fac929
+KAKAO_CALLBACK_URL=http://localhost:3000/auth/kakao/callback
