--- src/modules/auth/strategies/LocalStrategy.ts
@@ -1,19 +1,0 @@
-import { Strategy } from 'passport-local';
-import { PassportStrategy } from '@nestjs/passport';
-import { Injectable, UnauthorizedException } from '@nestjs/common';
-import { AuthService } from '../AuthService';
-
-@Injectable()
-export class LocalStrategy extends PassportStrategy(Strategy) {
-    constructor(private authService: AuthService) {
-        super();
-    }
-
-    async validate(userId: number, password: string): Promise<any> {
-        const user = await this.authService.validateUser(userId, password);
-        if (!user) {
-            throw new UnauthorizedException();
-        }
-        return user;
-    }
-}
