+++ src/modules/users/auth/AuthController.ts
@@ -1,18 +1,11 @@
-import { Controller, Get, HttpCode, HttpStatus, Post, Req, UseGuards } from '@nestjs/common';
+import { Controller, Get, HttpCode, Req, UseGuards } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
 import { AuthService } from './AuthService';
-import { UserKakaoDto } from '@/modules/auth/dtos';
+import { UserKakaoDto } from '@/modules/users/auth/dtos';
 
 @Controller('auth')
 export class AuthController {
     constructor(private readonly authService: AuthService) {}
-
-    /*
-    @UseGuards(LocalAuthGuard)
-    @Post('/login')
-    async login(@Req() req) {
-        return req.user;
-    }*/
 
     @Get('/kakao')
     @UseGuards(AuthGuard('kakao'))
