+++ src/modules/users/auth/strategies/index.ts
@@ -1,2 +1,1 @@
-export * from './LocalStrategy';
 export * from './KakaoStrategy';
